V56-82 | Connect gemini to database quering on schema and readonly queries
